feat: add login hook
